requirements.txt
streamlit
tensorflow
numpy
pillow
opencv-python